fix: add PartialOrd to ProtocolVersion

+++ questdb-rs/build.rs
@@ -167,7 +167,7 @@ pub mod json_tests {
             fn test_{:03}_{}(
                 #[values(ProtocolVersion::V1, ProtocolVersion::V2, ProtocolVersion::V3)] version: ProtocolVersion
             ) -> TestResult {{
-                if (version as u8) < {} {{
+                if version < {} {{
                     return Ok(());
                 }}
                 let mut buffer = Buffer::new(version);
